Move scanner processors to a common package
