+++ vulnerability-analyzer/src/main/java/org/hyades/VulnerabilityAnalyzerTopology.java
@@ -22,10 +22,10 @@ import org.hyades.common.KafkaTopic;
 import org.hyades.config.InternalScannerConfig;
 import org.hyades.config.OssIndexConfig;
 import org.hyades.config.SnykConfig;
-import org.hyades.processor.internal.InternalScannerProcessorSupplier;
+import org.hyades.processor.scanner.internal.InternalScannerProcessorSupplier;
 import org.hyades.processor.misc.TombstoneEmittingProcessorSupplier;
-import org.hyades.processor.ossindex.OssIndexProcessorSupplier;
-import org.hyades.processor.snyk.SnykProcessorSupplier;
+import org.hyades.processor.scanner.ossindex.OssIndexProcessorSupplier;
+import org.hyades.processor.scanner.snyk.SnykProcessorSupplier;
 import org.hyades.proto.KafkaProtobufSerde;
 import org.hyades.proto.vulnanalysis.internal.v1beta1.ExpectedScans;
 import org.hyades.proto.vulnanalysis.internal.v1beta1.ScanTask;
