+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/internal/InternalScannerProcessor.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.internal;
+package org.hyades.processor.scanner.internal;
 
 import com.github.packageurl.MalformedPackageURLException;
 import com.github.packageurl.PackageURL;
