+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/internal/InternalScannerProcessorSupplier.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.internal;
+package org.hyades.processor.scanner.internal;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.quarkus.cache.Cache;
