+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/ossindex/OssIndexProcessor.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.ossindex;
+package org.hyades.processor.scanner.ossindex;
 
 import com.github.packageurl.PackageURL;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
