+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/ossindex/OssIndexProcessorConfiguration.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.snyk;
+package org.hyades.processor.scanner.ossindex;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
@@ -13,8 +13,8 @@ import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.state.KeyValueStore;
 import org.apache.kafka.streams.state.StoreBuilder;
 import org.hyades.config.CircuitBreakerConfig;
+import org.hyades.config.OssIndexConfig;
 import org.hyades.config.RetryConfig;
-import org.hyades.config.SnykConfig;
 import org.hyades.processor.retry.RetryableRecord;
 import org.hyades.proto.vulnanalysis.internal.v1beta1.ScanTask;
 import org.hyades.proto.vulnanalysis.v1.ScanKey;
@@ -27,13 +27,13 @@ import javax.ws.rs.WebApplicationException;
 import static org.apache.kafka.streams.state.Stores.inMemoryKeyValueStore;
 import static org.apache.kafka.streams.state.Stores.keyValueStoreBuilder;
 
-class SnykProcessorConfiguration {
+class OssIndexProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("snykBatchStoreBuilder")
+    @Named("ossIndexBatchStoreBuilder")
     StoreBuilder<KeyValueStore<ScanKey, RetryableRecord<String, ScanTask>>> batchStoreBuilder() {
-        return keyValueStoreBuilder(inMemoryKeyValueStore("snyk-batch-store"),
+        return keyValueStoreBuilder(inMemoryKeyValueStore("ossindex-batch-store"),
                 new ObjectMapperSerde<>(ScanKey.class),
                 Serdes.serdeFrom(new ObjectMapperSerializer<>(),
                         new ObjectMapperDeserializer<>(new TypeReference<>() {
@@ -42,9 +42,9 @@ class SnykProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("snykRetryStoreBuilder")
+    @Named("ossIndexRetryStoreBuilder")
     StoreBuilder<KeyValueStore<ScanKey, RetryableRecord<String, ScanTask>>> retryStoreBuilder() {
-        return keyValueStoreBuilder(inMemoryKeyValueStore("snyk-retry-store"),
+        return keyValueStoreBuilder(inMemoryKeyValueStore("ossindex-retry-store"),
                 new ObjectMapperSerde<>(ScanKey.class),
                 Serdes.serdeFrom(new ObjectMapperSerializer<>(),
                         new ObjectMapperDeserializer<>(new TypeReference<>() {
@@ -53,8 +53,8 @@ class SnykProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("snykCircuitBreaker")
-    CircuitBreaker circuitBreaker(final SnykConfig config,
+    @Named("ossIndexCircuitBreaker")
+    CircuitBreaker circuitBreaker(final OssIndexConfig config,
                                   final MeterRegistry meterRegistry) {
         final CircuitBreakerRegistry registry = CircuitBreakerRegistry.of(
                 CircuitBreakerConfig.toConfigBuilder(config.circuitBreaker())
@@ -66,13 +66,13 @@ class SnykProcessorConfiguration {
                 .ofCircuitBreakerRegistry(registry)
                 .bindTo(meterRegistry);
 
-        return registry.circuitBreaker("snyk");
+        return registry.circuitBreaker("ossindex");
     }
 
     @Produces
     @ApplicationScoped
-    @Named("snykRetryIntervalFunction")
-    IntervalFunction retryIntervalFunction(final SnykConfig config) {
+    @Named("ossIndexRetryIntervalFunction")
+    IntervalFunction retryIntervalFunction(final OssIndexConfig config) {
         return RetryConfig.toIntervalFunction(config.retry());
     }
