+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/ossindex/OssIndexProcessorSupplier.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.ossindex;
+package org.hyades.processor.scanner.ossindex;
 
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.core.IntervalFunction;
