+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/snyk/SnykProcessor.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.snyk;
+package org.hyades.processor.scanner.snyk;
 
 import com.github.packageurl.MalformedPackageURLException;
 import com.github.packageurl.PackageURL;
