+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/snyk/SnykProcessorConfiguration.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.ossindex;
+package org.hyades.processor.scanner.snyk;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
@@ -13,8 +13,8 @@ import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.streams.state.KeyValueStore;
 import org.apache.kafka.streams.state.StoreBuilder;
 import org.hyades.config.CircuitBreakerConfig;
-import org.hyades.config.OssIndexConfig;
 import org.hyades.config.RetryConfig;
+import org.hyades.config.SnykConfig;
 import org.hyades.processor.retry.RetryableRecord;
 import org.hyades.proto.vulnanalysis.internal.v1beta1.ScanTask;
 import org.hyades.proto.vulnanalysis.v1.ScanKey;
@@ -27,13 +27,13 @@ import javax.ws.rs.WebApplicationException;
 import static org.apache.kafka.streams.state.Stores.inMemoryKeyValueStore;
 import static org.apache.kafka.streams.state.Stores.keyValueStoreBuilder;
 
-class OssIndexProcessorConfiguration {
+class SnykProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("ossIndexBatchStoreBuilder")
+    @Named("snykBatchStoreBuilder")
     StoreBuilder<KeyValueStore<ScanKey, RetryableRecord<String, ScanTask>>> batchStoreBuilder() {
-        return keyValueStoreBuilder(inMemoryKeyValueStore("ossindex-batch-store"),
+        return keyValueStoreBuilder(inMemoryKeyValueStore("snyk-batch-store"),
                 new ObjectMapperSerde<>(ScanKey.class),
                 Serdes.serdeFrom(new ObjectMapperSerializer<>(),
                         new ObjectMapperDeserializer<>(new TypeReference<>() {
@@ -42,9 +42,9 @@ class OssIndexProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("ossIndexRetryStoreBuilder")
+    @Named("snykRetryStoreBuilder")
     StoreBuilder<KeyValueStore<ScanKey, RetryableRecord<String, ScanTask>>> retryStoreBuilder() {
-        return keyValueStoreBuilder(inMemoryKeyValueStore("ossindex-retry-store"),
+        return keyValueStoreBuilder(inMemoryKeyValueStore("snyk-retry-store"),
                 new ObjectMapperSerde<>(ScanKey.class),
                 Serdes.serdeFrom(new ObjectMapperSerializer<>(),
                         new ObjectMapperDeserializer<>(new TypeReference<>() {
@@ -53,8 +53,8 @@ class OssIndexProcessorConfiguration {
 
     @Produces
     @ApplicationScoped
-    @Named("ossIndexCircuitBreaker")
-    CircuitBreaker circuitBreaker(final OssIndexConfig config,
+    @Named("snykCircuitBreaker")
+    CircuitBreaker circuitBreaker(final SnykConfig config,
                                   final MeterRegistry meterRegistry) {
         final CircuitBreakerRegistry registry = CircuitBreakerRegistry.of(
                 CircuitBreakerConfig.toConfigBuilder(config.circuitBreaker())
@@ -66,13 +66,13 @@ class OssIndexProcessorConfiguration {
                 .ofCircuitBreakerRegistry(registry)
                 .bindTo(meterRegistry);
 
-        return registry.circuitBreaker("ossindex");
+        return registry.circuitBreaker("snyk");
     }
 
     @Produces
     @ApplicationScoped
-    @Named("ossIndexRetryIntervalFunction")
-    IntervalFunction retryIntervalFunction(final OssIndexConfig config) {
+    @Named("snykRetryIntervalFunction")
+    IntervalFunction retryIntervalFunction(final SnykConfig config) {
         return RetryConfig.toIntervalFunction(config.retry());
     }
