+++ vulnerability-analyzer/src/main/java/org/hyades/processor/scanner/snyk/SnykProcessorSupplier.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.snyk;
+package org.hyades.processor.scanner.snyk;
 
 import io.github.resilience4j.circuitbreaker.CircuitBreaker;
 import io.github.resilience4j.core.IntervalFunction;
