+++ vulnerability-analyzer/src/test/java/org/hyades/VulnerabilityAnalyzerTopologyTest.java
@@ -13,9 +13,9 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.streams.processor.api.ContextualProcessor;
 import org.apache.kafka.streams.processor.api.Record;
 import org.hyades.common.KafkaTopic;
-import org.hyades.processor.internal.InternalScannerProcessorSupplier;
-import org.hyades.processor.ossindex.OssIndexProcessorSupplier;
-import org.hyades.processor.snyk.SnykProcessorSupplier;
+import org.hyades.processor.scanner.internal.InternalScannerProcessorSupplier;
+import org.hyades.processor.scanner.ossindex.OssIndexProcessorSupplier;
+import org.hyades.processor.scanner.snyk.SnykProcessorSupplier;
 import org.hyades.proto.KafkaProtobufSerde;
 import org.hyades.proto.vulnanalysis.internal.v1beta1.ScanTask;
 import org.hyades.proto.vulnanalysis.v1.Component;
