+++ vulnerability-analyzer/src/test/java/org/hyades/processor/scanner/internal/InternalScannerProcessorTest.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.internal;
+package org.hyades.processor.scanner.internal;
 
 import io.quarkus.cache.Cache;
 import io.quarkus.cache.CacheName;
@@ -14,6 +14,8 @@ import org.apache.kafka.streams.TopologyTestDriver;
 import org.apache.kafka.streams.test.TestRecord;
 import org.hyades.VulnerabilityAnalyzerTestProfile;
 import org.hyades.model.VulnerableSoftware;
+import org.hyades.processor.scanner.internal.InternalScannerProcessor;
+import org.hyades.processor.scanner.internal.InternalScannerProcessorSupplier;
 import org.hyades.proto.KafkaProtobufDeserializer;
 import org.hyades.proto.KafkaProtobufSerializer;
 import org.hyades.proto.vuln.v1.Source;
