+++ vulnerability-analyzer/src/test/java/org/hyades/processor/scanner/ossindex/OssIndexProcessorTest.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.ossindex;
+package org.hyades.processor.scanner.ossindex;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,6 +23,7 @@ import org.hyades.VulnerabilityAnalyzerTestProfile;
 import org.hyades.client.ossindex.ComponentReport;
 import org.hyades.client.ossindex.ComponentReportRequest;
 import org.hyades.client.ossindex.OssIndexClient;
+import org.hyades.processor.scanner.ossindex.OssIndexProcessorSupplier;
 import org.hyades.proto.KafkaProtobufDeserializer;
 import org.hyades.proto.KafkaProtobufSerializer;
 import org.hyades.proto.vuln.v1.Source;
