+++ vulnerability-analyzer/src/test/java/org/hyades/processor/scanner/snyk/SnykProcessorTest.java
@@ -1,4 +1,4 @@
-package org.hyades.processor.snyk;
+package org.hyades.processor.scanner.snyk;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
